requirements.txt
google-analytics-data>=0.18.0
pyyaml>=6.0
